Restructured EvaluationFunction to enable future GA
