+++ src/AI/Genetic_Algorithm/GeneticAlg_Eval.java
@@ -1,0 +1,20 @@
+package AI.Genetic_Algorithm;
+
+import AI.AI;
+
+public class GeneticAlg_Eval {
+
+    public static void main(String[] args) {
+//        START
+//        Generate the initial population
+//        Compute fitness
+//        REPEAT
+//                Selection
+//        Crossover
+//                Mutation
+//        Compute fitness
+//        UNTIL population has converged
+//        STOP
+
+    }
+}
